Move TabNavigator into its own file
